project/requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
librosa>=0.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
soundfile>=0.12.0
openai-whisper>=20230918
